backend/requirements.txt
asgiref==3.2.3
django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
